+++ src/native.c
@@ -5,18 +5,19 @@
 #include "include/interpreter.h"
 #include "include/object.h"
 #include "include/utils.h"
+#include <stddef.h>
 #include <stdint.h>
 #include <stdio.h>
 #include <time.h>
 
-static PValue ntvClock(PInterpreter *it, PValue *args, int argc) {
+static PValue ntvClock(PInterpreter *it, PValue *args, size_t argc) {
     clock_t c = clock();
     double sec = (double)c / (double)CLOCKS_PER_SEC;
     return MakeNumber(sec);
 }
 
-static PValue ntvShow(PInterpreter *it, PValue *args, int argc) {
-    for (int i = 0; i < argc; i++) {
+static PValue ntvShow(PInterpreter *it, PValue *args, size_t argc) {
+    for (size_t i = 0; i < argc; i++) {
         PrintValue(&args[i]);
         if (i + 1 < argc) {
             printf(" ");
